Partner info added to course purchase

+++ app/src/main/java/com/engineersapps/eapps/models/transactions/PayInvoiceResponse.kt
@@ -15,7 +15,7 @@ data class Salesinvoice(val udid: String?, val createdAt: String?, val updatedAt
 
 data class CreateOrderBody(var StudentID: Int = 0, var StudentMobile: String = "",
                            var GrandTotal: Int = 0, var PaidAmount: Int = 0,
-                           var DueAmount: Int = 0, var discount: Int = 0,
+                           var DueAmount: Int = 0, var discount: Double = 0.0,
                            var institute: String = "", var upazila: String = "",
                            var city: String = "", var UpazilaID: Int = 0,
                            var CityID: Int = 0, var InvoiceID: String = "",
